move endpoint to internal controller
